refactor(graph): extract GraphPersistence into module and wire restore with deps
